# services/memory-api/openapi.json
{"openapi":"3.1.0","info":{"title":"memory-api","description":"Read-only REST mirror of the graph-query MCP tools. Loads NetworkX graphs once at startup and serves search, neighbor, path, community, and subgraph queries over HTTP. Bearer-token auth on every endpoint except /healthz.","version":"0.1.0"},"paths":{"/healthz":{"get":{"tags":["meta"],"summary":"Healthz","operationId":"healthz_healthz_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthResponse"}}}}}}},"/search":{"get":{"tags":["query"],"summary":"Search","operationId":"search_search_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"query","in":"query","required":true,"schema":{"type":"string","minLength":1,"title":"Query"}},{"name":"scope","in":"query","required":false,"schema":{"type":"string","default":"personal","title":"Scope"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":200,"minimum":1,"default":20,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/NodeSummary"},"title":"Response Search Search Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/node/{node_id}":{"get":{"tags":["query"],"summary":"Node Info","operationId":"node_info_node__node_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"node_id","in":"path","required":true,"schema":{"type":"string","title":"Node Id"}},{"name":"scope","in":"query","required":false,"schema":{"type":"string","default":"personal","title":"Scope"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NodeInfo"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/neighbors/{node_id}":{"get":{"tags":["query"],"summary":"Neighbors","operationId":"neighbors_neighbors__node_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"node_id","in":"path","required":true,"schema":{"type":"string","title":"Node Id"}},{"name":"scope","in":"query","required":false,"schema":{"type":"string","default":"personal","title":"Scope"}},{"name":"max_hops","in":"query","required":false,"schema":{"type":"integer","maximum":4,"minimum":1,"default":1,"title":"Max Hops"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":500,"minimum":1,"default":30,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/NodeSummaryWithHops"},"title":"Response Neighbors Neighbors  Node Id  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/path":{"get":{"tags":["query"],"summary":"Path","operationId":"path_path_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"source","in":"query","required":true,"schema":{"type":"string","title":"Source"}},{"name":"target","in":"query","required":true,"schema":{"type":"string","title":"Target"}},{"name":"scope","in":"query","required":false,"schema":{"type":"string","default":"personal","title":"Scope"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PathResult"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/community/{node_id}":{"get":{"tags":["query"],"summary":"Community","operationId":"community_community__node_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"node_id","in":"path","required":true,"schema":{"type":"string","title":"Node Id"}},{"name":"scope","in":"query","required":false,"schema":{"type":"string","default":"personal","title":"Scope"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":500,"minimum":1,"default":30,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/NodeSummary"},"title":"Response Community Community  Node Id  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/subgraph":{"post":{"tags":["query"],"summary":"Subgraph","operationId":"subgraph_subgraph_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SubgraphRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SubgraphResult"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/top-nodes":{"get":{"tags":["query"],"summary":"Top Nodes","operationId":"top_nodes_top_nodes_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"n","in":"query","required":false,"schema":{"type":"integer","maximum":500,"minimum":1,"default":20,"title":"N"}},{"name":"scope","in":"query","required":false,"schema":{"type":"string","default":"personal","title":"Scope"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/NodeSummary"},"title":"Response Top Nodes Top Nodes Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"HealthResponse":{"properties":{"status":{"type":"string","title":"Status"},"graphs_loaded":{"items":{"type":"string"},"type":"array","title":"Graphs Loaded"}},"type":"object","required":["status","graphs_loaded"],"title":"HealthResponse"},"NodeInfo":{"properties":{"id":{"type":"string","title":"Id"},"attributes":{"additionalProperties":true,"type":"object","title":"Attributes"},"degree":{"type":"integer","title":"Degree"},"top_neighbors":{"items":{"$ref":"#/components/schemas/NodeSummary"},"type":"array","title":"Top Neighbors"}},"type":"object","required":["id","attributes","degree","top_neighbors"],"title":"NodeInfo"},"NodeSummary":{"properties":{"id":{"type":"string","title":"Id"},"name":{"type":"string","title":"Name"},"degree":{"type":"integer","title":"Degree"},"community":{"anyOf":[{},{"type":"null"}],"title":"Community"}},"type":"object","required":["id","name","degree"],"title":"NodeSummary"},"NodeSummaryWithHops":{"properties":{"id":{"type":"string","title":"Id"},"name":{"type":"string","title":"Name"},"degree":{"type":"integer","title":"Degree"},"community":{"anyOf":[{},{"type":"null"}],"title":"Community"},"hops":{"type":"integer","title":"Hops"}},"type":"object","required":["id","name","degree","hops"],"title":"NodeSummaryWithHops"},"PathResult":{"properties":{"source":{"type":"string","title":"Source"},"target":{"type":"string","title":"Target"},"path":{"items":{"$ref":"#/components/schemas/NodeSummary"},"type":"array","title":"Path"},"length":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Length"}},"type":"object","required":["source","target","path","length"],"title":"PathResult"},"SubgraphRequest":{"properties":{"concepts":{"items":{"type":"string"},"type":"array","minItems":1,"title":"Concepts"},"max_hops":{"type":"integer","maximum":4.0,"minimum":1.0,"title":"Max Hops","default":1},"limit":{"type":"integer","maximum":500.0,"minimum":1.0,"title":"Limit","default":40},"scope":{"type":"string","title":"Scope","default":"personal"}},"type":"object","required":["concepts"],"title":"SubgraphRequest"},"SubgraphResult":{"properties":{"seeds":{"items":{"type":"string"},"type":"array","title":"Seeds"},"missing":{"items":{"type":"string"},"type":"array","title":"Missing"},"nodes":{"items":{"$ref":"#/components/schemas/NodeSummary"},"type":"array","title":"Nodes"},"edges":{"items":{"additionalProperties":{"type":"string"},"type":"object"},"type":"array","title":"Edges"},"node_count":{"type":"integer","title":"Node Count"},"edge_count":{"type":"integer","title":"Edge Count"}},"type":"object","required":["seeds","missing","nodes","edges","node_count","edge_count"],"title":"SubgraphResult"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}